# docs/swagger.yaml
openapi: '3.0.0'
info:
  title: Shoplist api service
  version: '1.0'
servers:
  - url: https://api.server.test/v1
security:
  - token: []
paths:
...
  "/addItem": 
    post:
      operationId: addItem
      requestBody:
        $ref: "#/components/requestBodies/Item_request"
      tags:
        - Shopping items
      summary: Добавление товара в покупку
      responses:
        "200":
          $ref: "#/components/responses/Item_200"
        "400":
          $ref: "#/components/responses/Item_400"
        "401":
          $ref: "#/components/responses/Base_401"
        "404":
          $ref: "#/components/responses/Base_404"  
        "405":
          $ref: "#/components/responses/Base_405"
        "500":
          $ref: "#/components/responses/Base_500"
  "/deleteItems":
    post:
      operationId: "deleteItems"
      requestBody:
        $ref: "#/components/requestBodies/Delete_items_request"
      tags:
        - Shopping items
      summary: Удаление товаров
      responses:
        "200":
          $ref: "#/components/responses/Base_200"
        "400":
          $ref: "#/components/responses/Base_400"
        "401":
          $ref: "#/components/responses/Base_401"
        "405":
          $ref: "#/components/responses/Base_405"
        "500":
          $ref: "#/components/responses/Base_500"
...
  "/users":
    get:
      operationId: "getUsers"
      parameters:
        - $ref: "#/components/parameters/telegram_user_id"
        - $ref: "#/components/parameters/comunity_id"
      tags:
        - Users
      summary: Получение юзера по telegram user id
      responses:
        "200":
          $ref: "#/components/responses/Users_200"
        "400":
          $ref: "#/components/responses/Base_400"
        "401":
          $ref: "#/components/responses/Base_401"
        "404":
          $ref: "#/components/responses/Base_404"
        "405":
          $ref: "#/components/responses/Base_405"
        "500":
          $ref: "#/components/responses/Base_500"
    post:
      operationId: "createUser"
      requestBody:
        $ref: "#/components/requestBodies/User_request"
      tags:
        - Users
      summary: Добавление юзера
      responses:
        "200":
          $ref: "#/components/responses/Users_200"
        "400":
          $ref: "#/components/responses/Base_400"
        "401":
          $ref: "#/components/responses/Base_401"
        "405":
          $ref: "#/components/responses/Base_405"
        "500":
          $ref: "#/components/responses/Base_500"
    patch:
      operationId: "updateUser"
      parameters:
        - $ref: "#/components/parameters/user_id"
      requestBody:
        $ref: "#/components/requestBodies/User_request"
      tags:
        - Users
      summary: Добавление юзера
      responses:
        "200":
          $ref: "#/components/responses/Base_200"
        "400":
          $ref: "#/components/responses/Base_400"
        "401":
          $ref: "#/components/responses/Base_401"
        "404":
          $ref: "#/components/responses/Base_404"
        "405":
          $ref: "#/components/responses/Base_405"
        "500":
          $ref: "#/components/responses/Base_500"
components:
  securitySchemes:
    accesstoken:
      type: apiKey
      in: query
      name: accesstoken
  schemas:
      Base:
        description: Возвращает статус ответа
        type: object
        properties:
          version:
            type: string
            description: Вeрсия ответа сервиса
            example: 1.0.0
      Success:
        description: Ответ успешно получен
        allOf:
          - $ref: "#/components/schemas/Base"
          - type: object
            title: Success
            properties:
              message:
                type: string
                description: Сообщение ответа
                example: success
              errors:
                type: array
                items: {}
                description: Ошибки ответа
            required:
              - message
              - errors
      Error:
        description: Ошибка
        allOf:
          - $ref: "#/components/schemas/Base"
          - type: object
            title: Error
            properties:
              data:
                example: null
            required:
              - data
      Error_400:
        allOf:
          - $ref: "#/components/schemas/Error"
          - type: object
            title: Error_400
            properties:
              message:
                type: string
                example: Validation error
              errors:
                type: array
                items: {}
            required:
              - message
              - errors
      Error_401:
        allOf:
          - $ref: "#/components/schemas/Error"
          - type: object
            title: Error_401
            properties:
              message:
                type: string
                example: Unauthorized
              errors:
                type: array
                items: {}
            required:
              - message
              - errors
      Error_404:
        allOf:
          - $ref: "#/components/schemas/Error"
          - type: object
            title: Error_404
            properties:
              message:
                type: string
                description: сообщение об ошибке
                example: Entity not found
              errors:
                type: array
                items: {}
                example: null
            required: 
              - message
              - errors
      Error_405:
        allOf:
          - $ref: "#/components/schemas/Error"
          - type: object
            title: Error_405
            properties:
              message:
                type: string
                description: сообщение ответа
                example: Method ${method} not allowed
              errors:
                type: array
                items: {}
                example: null
      Error_500:
        allOf:
          - $ref: "#/components/schemas/Error"
          - type: object
            title: Error_500
            properties:
              message:
                type: string
                description: сообщение ответа
                example: Internal server error
              errors:
                example: null
            required:
              - message
              - errors
      shoppingItemParams:
        title: Параметры товара в покупке
        properties:
          productName:
            type: string
            description: Наименование товара
            example: Хлеб
            minLength: 1
          quantity:
            type: integer
            description: Количество товара
            example: 1
          categoryID:
            type: integer
            description: Идентификатор категории товара
            example: 5
          complete:
            type: boolean
            description: Признак выполнения
            example: true
          listID:
            type: integer
            description: Идентификатор списка покупок
            example: 5
        required:
          - productName
          - quantity
          - categoryID
          - complete
          - listID
      user:
        title: Данные юзера
        properties:
          telegram_id:
            type: integer
            description: telegram_id
            example: 364534
          telegram_username:
            type: string
            description: telegram_username
            example: jhon_snow
          comunity_id:
            type: string
            description: comunity_id
            example: "gfrf4jfdf"
          token:
            type: string
            description: token
            example: "gfrf4jfdf"
          chat_id:
            type: integer
            format: int64
            description: chat_id
            example: 324432
      userWithID:
        allOf:
          - type: object
            properties:
              id:
                type: integer
                description: Идентификатор юзера
                example: 55
          - $ref: "#/components/schemas/user"        
      shoppingItemParamsWithId:
        allOf:
          - type: object
            properties:
              id:
                type: integer
                description: Идентификатор товара
                example: 55
          - $ref: "#/components/schemas/shoppingItemParams"
          - type: object
            required:
              - id
              - productName
              - quantity
              - categoryID
              - complete
              - listID
...
      Delete_ids:
        title: Идентификаторы для удаления
        properties:
          ids:
            type: array
            description: Массив идентификаторв для удаления
            items:
              type: integer
        required:
          - ids
...
  requestBodies:
    Item_request:
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/shoppingItemParams"
      required: true
...
    Delete_items_request:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Delete_ids"
    User_request:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/user"
  responses:
      Base_200:
        description: Success
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Success"
      Base_400:
        description: Bad request
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_400"
      Base_401:
        description: Unauthorized
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_401"
      Base_404:
        description: Not found
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_404"
      Base_405:
        description: Invalid method
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_405"
      Base_500:
        description: Internal server error
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_500"
...
      Item_200:
        description: Добавленный товар
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Success"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: "#/components/schemas/shoppingItemParamsWithId"
      Item_400:
        description: Validation error
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Error_400"
                - type: object
                  properties:
                    errors: 
                      type: object
                      properties:
                        validation:
                          type: object
                          properties:
                            productName:
                              type: string
                              example: "min"
                              enum:
                                - min
                            quantity:
                              type: string
                              description: Количество товара
                              example: "format"
                              enum:
                                - format
                            categoryID:
                              type: string
                              description: Идентификатор категории товара
                              example: "format"
                              enum:
                                - format
                            listID:
                              type: string
                              description: Идентификатор списка покупок
                              example: "format"
                              enum:
                                - format 
...
      Users_200:
        description: Данные юзера
        content: 
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Success"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: "#/components/schemas/userWithID"
  parameters:
...
    telegram_user_id:
      name: telegram_user_id
      in: query
      description: telegram user id
      schema:
        type: integer
    comunity_id:
      name: comunity_id
      in: query
      description: comunity_id
      schema:
        type: string
    user_id:
      name: user_id
      in: query
      description: user_id
      required: true
      schema:
        type: integer
...